geoip/geoip/PyGeoIP/Geo/requirements.txt
python-geoip
python-geoip-geolite2
djangorestframework
redis
pytz
requests
